third_party/blink/web_tests/external/wpt/css/css-overflow/inheritance-expected.txt

This is a testharness.js-based test.
[FAIL] Property block-ellipsis has initial value none
  assert_true: block-ellipsis doesn't seem to be supported in the computed style expected true got false
[FAIL] Property block-ellipsis inherits
  assert_true: block-ellipsis doesn't seem to be supported in the computed style expected true got false
[FAIL] Property continue has initial value auto
  assert_true: continue doesn't seem to be supported in the computed style expected true got false
[FAIL] Property continue does not inherit
  assert_true: expected true got false
[FAIL] Property max-lines has initial value none
  assert_true: max-lines doesn't seem to be supported in the computed style expected true got false
[FAIL] Property max-lines does not inherit
  assert_true: expected true got false
Harness: the test ran to completion.
